move more pages over
