+++ apps/temp-docs/pages/index.mdx
@@ -1,7 +1,9 @@
----
-title: Introduction
-description: Introduction to Supabase
----
+import Layout from '~/layouts/Guide'
+
+export const meta = {
+  title: 'Introduction',
+  description: 'Introduction to Supabase',
+}
 
 # Introduction
 
@@ -43,3 +45,5 @@ The documentation is divided into 3 sections.
   <LinkCard title="Overview" description="A summary of the tools and how to get started" />
   <LinkCard title="Overview" description="A summary of the tools and how to get started" />
 </LinkCardsWrapper>
+
+export default (context) => <Layout meta={meta} children={context.children} context={context} />
